Bug 1925134 [wpt PR 48655] - [text-decoration] Fix crashes for multiline ellipsis with `::first-line`, a=testonly
Automatic update from web-platform-tests
[text-decoration] Fix crashes for multiline ellipsis with `::first-line`

The ellipsis should be styled according to the block[1]. It
was using the first-line style of the block, because it's the
common case, and it's the same style unless `::first-line` is
applied.

This patch changes it to distinguish ellipsis on the first
line from other cases.

Wrong style can cause wrong rendering, but when
`text-decoration` is involved, it can confuse decoration box
calculation and result in crashes. This patch fixes both the
crashes and the correctness.

[1] https://www.w3.org/TR/css-overflow-3/#ellipsing-details

Fixed: 358957652
Change-Id: I26af0ae034876bdc0b0cc114670c52d24e2f36a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5936973
Commit-Queue: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1369664}

--

wpt-commits: 1c795222d59595eadcd348faf86334ae6877224c
wpt-pr: 48655