Bug 1926359 - Add tab group line to dragged tabs when dragging into a group. r=dwalker,tabbrowser-reviewers
Ideally the line should cover the whole group, but I don't know that we can do this given our use of display:contents. If someone has ideas we can file a followup on that. As it stands the line will extend under dragged tabs themselves, but left or right from those there can be gaps in the line.

Differential Revision: https://phabricator.services.mozilla.com/D226634