Bug 1926584 - Fix menubar autohide detection. r=mac-reviewers,mstange
This fixes the autohide detection during the fullscreen transition. With
bug 1921819 reverted, this gets the right behavior on all configurations
I tested.

We need to directly read the user setting. Something a bit more elegant
which I tried but doesn't work would be to check the
NSApplicationPresentationOptions when entering fullscreen. However that
doesn't change at all.

Differential Revision: https://phabricator.services.mozilla.com/D226673