Bug 1927020 - Add/use new thread agnostic shutdown watcher for WebCodecs DOM objects. r=media-playback-reviewers,padenot
Because outstanding asynchronous promises may keep various WebCodecs
objects alive even if the explicit reference to the DOM object has gone
away in JS, we need a mechanism to trigger a teardown if we are still
waiting for those promises to be fulfilled come shutdown.

Unfortunately, the present mechanism holds onto a strong reference
itself, preventing cycle collection from ever freeing us, because it
isn't aware of this strong reference.

This patch adds a new pair of classes, media::ShutdownConsumer and
media::ShutdownWatcher for that purpose. ShutdownConsumer is intended to
be subclasses by the DOM objects to provide a method to be called on
thread shutdown (either main or DOM worker). ShutdownWatcher is intended
to be held onto by the DOM object until it is cycle collected, and
listens for the appropriate shutdown event for its thread to force
destruction otherwise. ShutdownWatcher holds a weak pointer to the DOM
object, and thus the Destroy method must be called prior to the DOM
object being freed.

Differential Revision: https://phabricator.services.mozilla.com/D226866