Bug 1926418 - Rename UniFFI async terms, r=markh
I'm hopping to add support for async Rust calls, which could be
confusing if we keep calling our current calls "async". This commit
moves us towards a future where there are 3 types of scaffolding calls:

- Sync (Rust sync function being called synchronously from JS)
- Async (Rust async function being called asynchronously from JS)
- AsyncWrapper (Rust sync function that's wrapped so it can be called asynchronously from JS)

The async-wrapper behavior is now opt-in rather than the default and the
config keys have changed slightly.

Differential Revision: https://phabricator.services.mozilla.com/D226566