Bug 1915351 part 10 - make copy image test more reliable r=dlp-reviewers,handyman
This part of the test fails often on try. I'm not sure what's going on
(sometimes a file gets put on the clipboard as well), but the point of
this test is to test content analysis, not Copy Image Contents, so
just set the clipboard contents directly.

I get passing test-verify builds with this change.

Differential Revision: https://phabricator.services.mozilla.com/D227037