Bug 1916650 - [devtools] Adapt breadcrumb to High Contrast Mode. r=accessibility-frontend-reviewers,devtools-reviewers,nlapre,ochameau.
We explicitely set the color for the tag element so it's easier
to work around UA button styling (especially around hover color),
and we set High Contrast Mode specific colors.

The style of the breadcrumb items was updated so the selected
background doesn't cover the separator icon (which is rendered
in the breadcrumb button before pseudo element).

Differential Revision: https://phabricator.services.mozilla.com/D226671