Bug 1926151 - Remove Desktop and Public Desktop from folders a pinnable shortcut can exist in. r=bhearsum
We prioritize using `ITaskbarManager` now, which requires a shortcut exist in a location visible from `shell:appsfolder`. `IPinnedList3` can use shortcuts from Desktop and Public Desktop (and potentially other locations) but there's no reason to check for these cases since we have already created the shortcut for `ITaskbarManager`. The rare case of failing to create as shortcut for `ITaskbarManager` is not worth accounting for given `IPinnedList3` is not supported in newer versions of Windows.

Differential Revision: https://phabricator.services.mozilla.com/D227134