Bug 1754452 - [devtools] Select the hovered document when using the node picker on Web Extensions. r=devtools-reviewers,nchevobbe
Before, we used to have a unique target, using internal frame switching.
Now that we have one real target actor per document, we should rather select
each newly hovered target at the TargetCommand level, so that the inspector's
markup view will switch to this document.

We are applying this logic only to Web Extensions as all targets in a tab
will be a children of the top level one, so that node picking's hovering
will always be able to select the hovered node in the markup view.
We may later extand this logic to the browser toolbox, once it supports one target per WindowGlobal.

Differential Revision: https://phabricator.services.mozilla.com/D220596