Bug 1928012 - Modernize implementation of GetCommonFallbackFonts for Windows. r=jwatt
Mapping fallback fonts based on script code rather than hard-coded
blocks of Unicode codepoints is much more readable and maintainable.

We should periodically review what fonts Windows is shipping, and add
entries for new scripts as they become available. (Note that it is not
necessary for a font to be listed here in order to be used; global
fallback will still find it, if available. The mappings here just
provide a fast-path that tries to avoid the cost of a global search.)

Differential Revision: https://phabricator.services.mozilla.com/D227660