Bug 1927795 - Part 2: Add SinglyLinkedList::removeRange and iterFrom methods r=sfink
The idea is to implement ArenaList with SinglyLinkedList, which only uses a
single word of storage.

This adds a couple of new methods that we'll need on the list representation.

removeRange is kinda awkward but is needed for compacting where we will have to
remove a section of the list.

iterFrom is used for incremental sweeping.

prepend is used in sweeping where we currently implement this by with an append
and a move operation.

Differential Revision: https://phabricator.services.mozilla.com/D227250