Bug 1895174 - Block WRusr.dll in utility processes. r=bobowen,win-reviewers,gstoll
WRusr.dll injection from WebRoot causes a crash in the utility LPAC
process used for PlayReady because on some versions of Windows,
CallNamedPipeW does COM if it fails to open the pipe in an AppContainer
process. This does not affect other processes though, so blocking
injection in utility processes should be enough.

Differential Revision: https://phabricator.services.mozilla.com/D228185