Bug 1929003 - [devtools] Use native profiler to collect JS Traces from DevTools. r=dthayer
Stop manually crafting the profiler data by using SpiderMonkey hooks.
Instead, use the profiler native pipeline to do the collection.

This still relies on JSTracer class in order to trigger the record on next interaction or page load.
One caveat is that it starts the profiler from the content process, which prevents recording screenshots correctly.

Differential Revision: https://phabricator.services.mozilla.com/D227831