Bug 1929636: Improve content analysis paste-blocked and warn dialogs when agent is not connected r=dlp-reviewers,gstoll
When the DLP agent can't be reached, we enact default DLP behavior based on the
browser.contentanalysis.default_result pref.  That still shows result dialogs
for pastes, and when default-warn is enabled, and the messages used for those
when the agent *is* connected would confuse the user here.  This changes those
messages.

Differential Revision: https://phabricator.services.mozilla.com/D228233