Bug 1922506: Update base64 canvas3. r=tjr
This patch changes data:content/type to data:image/jpeg. Base64 contents are almost the same, I'm not sure what exactly changed, but I basically redownloaded the image from its source and base64'ed it in hopes that we no longer get random error events in canvasdata3 receipe. This patch also updates the size of the canvas to 400x225 in order to match the image dimensions.

Differential Revision: https://phabricator.services.mozilla.com/D226635