Bug 1929739 [wpt PR 49017] - [line-clamp] Fix DCHECK failure with hidden block-in-inlines, a=testonly
Automatic update from web-platform-tests
[line-clamp] Fix DCHECK failure with hidden block-in-inlines

When painting a hidden-from-paint block-in-inline (which can happen if
the block-in-inline is after the clamp point in a line-clamp
container), there was a DCHECK failure since
`BoxFragmentPainter::PaintBoxItem` was not expected to be called on
hidden for paint items. This is because, in all other cases, this call
is skipped if the item is hidden, but that was not the case for block
in inlines. This patch fixes this by adding a condition before the
call in
`BoxFragmentPainter::PaintBoxDecorationBackgroundForBlockInInline`.

Bug: 40336192
Change-Id: I9a7a4acbb9e5e1cb42d27865665a5ce37d9bfa57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5987970
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1379688}

--

wpt-commits: 7d50b283168dc2facae92e26126f8996ed02ecbe
wpt-pr: 49017