Bug 1929755: Add supplemental UIA attribute definitions for MinGW builds,r=Jamie
Some UIA attributes are not yet defined on the version of MinGW that we are
using to build and test Gecko. Eventually, an upgrade to MinGW will make these
definitions redundant. In the meantime, this shim allows us to keep implementing
UIA features without breaking builds. The attributes defined in the new header
come directly from the UIAutomationClient.h file, and match the UIA docs.

Differential Revision: https://phabricator.services.mozilla.com/D228366