Bug 1926357: Rework wasmFailValidateText and assertErrorMessage for clarity. r=rhunt
The message for wasmFailValidateText made it sound like the compile had
not failed, when in fact it just did not fail "as expected", i.e. in the
expected way. That message is now clearer.

The message printing for assertErrorMessage mixed up the order in a
confusing way that made it seem like the top-level user message was
actually the most proximal cause of the exception. This was weird and
has been reordered. The code has also been reworked for clarity, making
it obvious that no exception can slip through, and handling the common
mistake of forgetting to provide an error message. (It is at least a
common mistake for me.)

Differential Revision: https://phabricator.services.mozilla.com/D227682