Bug 1856311 - Fix verifyDoNotSaveCreditCardFromPromptTest UI test r=aaronmt,ohorvath
The UI test was flaky from time to time because the click on the "never save" button couldn't be performed due to the fact that the prompt wasn't displayed.
To overcome this issues I've added 1s wait in between the credit card filling actions.

The UI test successfully passed 150x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D228656