Bug 1811560 - Simplify persist search telemetry test - r=daisuke,urlbar-reviewers
For some reason on Linux, going back to the SERP can end up
going to the first page that was loaded, hence the Promise that
awaits the SERP is loaded time outs.

One workaround is to make the SERP the first page. I think this
is an acceptable change because the test is trying to ensure
returning to a SERP and making another search results in the
correct telemetry.

Differential Revision: https://phabricator.services.mozilla.com/D228520