Bug 1928666 part 2 - Remove Range::onTableDestroyed. r=jonco
`onTableDestroyed` iterates over all ranges and this is blocking us from doing
background finalization of `Map` and `Set` objects.

Note that `Range` no longer stores a pointer to the table and the methods instead
take the `Map` or `Set` object as argument. This makes invalidating the ranges less
valuable than it was before.

Differential Revision: https://phabricator.services.mozilla.com/D228979