Bug 1931336 - Fix aliasing issue of MObjectKeysLength. r=iain
Global Value Numbering happen to alias multiple calls to MObjectKeysLength
assuming they were called on the same object. The problem comes from the lack of
registered memory dependency, which let GVN assumed that this operation is sane.

This change fixes this issue by adding a dependency on the original instruction,
which remains in the graph as a recovered instruction.

Differential Revision: https://phabricator.services.mozilla.com/D229187