Bug 1930322 part 3: Fixes to browser_caching_actions.js test. r=morgan
As part of these tests, we need to wait for all specified DOM events to fire on a node.
We do this by setting up a Promise for each event, wrapping these in a single Promise which resolves when all events fire, storing this wrapping Promise on the Window object, and later (in a separate task) waiting on this wrapping Promise.
Previously, after any event fired, we set the Window object variable holding the wrapping Promise to null in order to clean it up.
This resulted in a race condition where if the first event fired before the second task ran to wait on the wrapping Promise, the second task would simply await null, meaning that it didn't wait at all.
The practical upshot is that we could potentially pass the tests even if the required DOM events never fired and remain blissfully unaware.
To fix this, we instead clear the wrapping Promise variable in the second task after waiting on it.
Without this fix, the test I add in the next patch would pass even without the bug being fixed.
In addition, to make debugging timeouts easier, we now log the id of the element being tested.

Differential Revision: https://phabricator.services.mozilla.com/D229299