Bug 1928112 - Add the find in page bar as possible anchor for the download complete panel r=android-reviewers,harrisono
Also solve a related layout race with the find in page bar first appearing with a 0 height
and saving it as an anchor but for later layout changes we would not reposition the download
panel since in the meantime the height was updated to a bigger value and we'd assume this
bigger value was already used to position the download panel.
Now we'll ignore remembering an anchor if it's not visible or has a 0 height which will
allow to react more reliably to layout changes.

Differential Revision: https://phabricator.services.mozilla.com/D228968