Bug 1928112 - Ensure the right translationY if the anchor changes during the animation r=android-reviewers,tchoh
There is a race between responding to layout changes and animating the download complete panel.
With the navbar enabled when changing device orientation the bottom toolbar cycles through
no toolbar, just the address bar, addressbar + navbar.
If the animation to show the panel starts while the toolbar layout is still updated then it
will set a wrong translationY for placing the download complete panel above the toolbar.
If the animation ends after the layout has settled then we will not update the anchor and the
right translationY again since other code this has already been set.
To overcome all this when the animation ends we will set the current toolbar height as the
translationY value need to correctly place the download complete panel above the toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D228967