Bug 1931448 [wpt PR 49188] - Enable ShadowRealm testing for atob/btoa, a=testonly
Automatic update from web-platform-tests
Enable ShadowRealm testing for atob/btoa

Adds the appropriate metadata for the existing tests for the functionality
of atob() and btoa() to additionally be run in a ShadowRealm.

We also have to use the test harness's fetch_json() to fetch the test data
instead of fetch(), which is not exposed in ShadowRealm.

--

wpt-commits: f4c5e40b830cb5811d9838a03b1e95e0c9d4323f
wpt-pr: 49188