Bug 1919347 - Support getComputedStyle() for named view transition pseudo elements if the style is not available. r=view-transitions-reviewers,emilio
First, we have to allow the functional parameter for named view
transition pseudo elements.

In this patch, we fix the case when `getComputedStyle()` resolve the style
lazily (i.e. call `Servo_ResolveStyleLazily()`) if its style is not available.
In this case, we have to make sure the matching function work.

We will make sure `aElement->GetPseudoElement(aPseudo)` work in the next
patch so we can retrieve the computed style from the pseudo elements
directly.

Differential Revision: https://phabricator.services.mozilla.com/D230354