Bug 1916649 - [devtools] Add colored circle in overflow badge when related scroll badge was clicked. r=devtools-reviewers,jdescottes.
When the scroll badge is clicked, we expand the tree to show all the element the overflow,
and set a background color on the nodes to identify them.
In High Contrast Mode, the background color can't be set, and so we fail to identify the
overflowing nodes.
This patch adds a small circle in the overflow badge, of the same color than the active
scroll badge so people can see the relation between the scroll badge and the overflow one.

Differential Revision: https://phabricator.services.mozilla.com/D230664